Handle regions with `ChildRegion` as entry
During the first iteration outlining, we could have regions with, as
entry, the `ChildRegionDescriptor` referring to another nested region.
This is expected, and during the actual `clift.loop` creation, this will
be replaced with an actual `mlir::Block` being materialized.

During outlining though, we need to take care of this situation, and in
particular to ignore the presence of this node, and outline all other
possible late entries.