Improve `LateEntry` computation
When electing the late entry nodes in a region, we only take into
consideration nodes whose predecessor in not in the current region, but
is instead in the parent region (but not further up).

This is necessary because a late entry edge may bypass an arbitrary
number of nested regions, so the outlining needs to be perfomed when we
actually need to clone nodes in the correct `ParentRegion`.