Use `MapVector` for `dfs` visits
Use a `MapVector` instead of a `DenseMap` to encapsule the state of the
`llvm::df` visit. In this way, the order of insertion of nodes in the
regions is deterministic, due to the fact that the `NodesBetween`
primitive finds the nodes always in the same order on the visit stack.