GraphTraits: add `Undirected` marker class
Add a marker class that is used to tell the graph iterator to iterate
over the graph in a graph defined `Undirected` ordering.

The idea is to treat the edges in a graph as if they were non-directer,
and we implement this by concatenating the successor and predecessor
iterator for the `child_begin`/`child_end` trait.